requirements.txt
streamlit
pytesseract
pdfplumber
pillow
transformers
torch
sentencepiece
spacy
